split Global Data into submodules (#27)
* get_current_blockhash and start test for get valid blockhashes

* get tests working

* updating versions fixed a segfault issue

* remove unnecessary .gitignore

* cleanup

* create ValidBlockhashes

* move ValidBlockhashes

* remove unused function

* create DailyDistributionData

* create DailyDistributionData

* Add needs_signer/needs_writable to account verification

* prep work/ bug fixes

* UserData try_from &AccountInfo

* add iterator to DailyDistrubutionData

* add get_sum_last_n_interests() to DailyDistributionData

* make UserData fields pub

* get interest

* get interest

* add transfer

* transfer from banks

* account name consistency

* fix minor bugs

* DailyDistributionData insert

* get_owed_comptokens integration test + bugfixes

* cleanup

* added waitForTransactionConfirmation

* added waitForTransactionConfirmation

* added quick error handling

* refactor getValidBlockHashes

* fix comments

* fix mispelled comment

* fix bad merge

* get correct interest

* create VerifiedAccountInfo

* switch to return VerifiedAccountInfo

* move daily_distribution_data into own file

* move ValidBlockhashes into own file

* fix clippy warnings

* move global_data.rs into its submodule folder

---------
