+++ src/comptoken_proof.rs
@@ -5,7 +5,7 @@ use spl_token_2022::solana_program::{
     pubkey::Pubkey,
 };
 
-use crate::global_data::ValidBlockhashes;
+use crate::global_data::valid_blockhashes::ValidBlockhashes;
 
 // ensure this remains consistent with comptoken_proof.js
 const MIN_NUM_ZEROED_BITS: u32 = 3; // TODO: replace with permanent value
@@ -16,7 +16,7 @@ pub fn verify_proof(proof: &ComptokenProof, valid_blockhashes: &ValidBlockhashes
     let equal_hash: bool = proof.generate_hash() == proof.hash;
     let valid_hash_is_fresh: bool = !valid_blockhashes.is_valid_blockhash_stale();
     // hash duplicate check is part of inserting
-    return leading_zeros && recent_blockhash && equal_hash && valid_hash_is_fresh;
+    leading_zeros && recent_blockhash && equal_hash && valid_hash_is_fresh
 }
 
 pub const VERIFY_DATA_SIZE: usize = HASH_BYTES + mem::size_of::<u64>() + HASH_BYTES;
